test(cmd): consolidate pri tests into pri_test.go
